fix(web): fast refresh
