+++ html/src/components/functions/useAuth.tsx
@@ -1,0 +1,10 @@
+import * as React from "react";
+import { AuthContext } from "../../auth.tsx";
+
+export function useAuth() {
+  const context = React.useContext(AuthContext)
+  if (!context) {
+    throw new Error('useAuth must be used within an AuthProvider')
+  }
+  return context
+}
